chore: EXC-1749: Consolidate scheduling logic No functional changes, just moving functions in one place.
