+++ rs/execution_environment/src/execution_environment.rs
@@ -15,6 +15,7 @@ use crate::{
     hypervisor::Hypervisor,
     ic00_permissions::Ic00MethodPermissions,
     metrics::{CallTreeMetrics, CallTreeMetricsImpl, IngressFilterMetrics},
+    RoundSchedule,
 };
 use candid::Encode;
 use ic_base_types::PrincipalId;
@@ -3225,7 +3226,7 @@ impl ExecutionEnvironment {
                     .task_queue
                     .replace_paused_with_aborted_task(aborted_task);
             }
-            canister.apply_priority_credit();
+            RoundSchedule::apply_priority_credit(canister);
             let canister_id = canister.canister_id();
             canister.system_state.apply_ingress_induction_cycles_debit(
                 canister_id,
